requirements.txt
gensim
oci
python-dotenv
scikit-learn
matplotlib
japanize_matplotlib
numpy
seaborn
cohere
pandas
Jinja2
umap-learn
adjustText
ipywidgets
h5py
protobuf
fugashi
unidic-lite
tabulate
gradio